Reporting Issue: Does not work at all with Safari
Issues are unfortunately disabled for this repository. Hence my issue as a pull request. It is definitely the case that this package does not work at all with Safari. Sorry, hence the adjustment in the README.

Demo in latest Chrome:
![Chrome](https://user-images.githubusercontent.com/640639/142586831-8c9ebc0f-69ac-4fae-a70b-56272aadc4c9.png)

Demo in Safari Version 11.0.3 (13604.5.6)
![Safari](https://user-images.githubusercontent.com/640639/142586848-3b147c56-022e-412d-b2b3-5202577430aa.png)

If Safari on MacOS is supported in a newer version though, then it would be very helpful to specify the minimum version in the README.






see also https://github.com/JohannesHoppe/angular-date-value-accessor/issues/22